requirements.txt
discord.py==2.6.3
python-dotenv
aiohttp
SpeechRecognition
pydub
ffmpeg-python
requests
pyjwt